eks/requirements.gateway.txt
grpcio==1.59.2 
tensorflow-serving-api==2.14.0
tensorflow==2.14.0
numpy==1.24.3
flask==3.0.3
gunicorn==23.0.0
pillow==11.0.0